Use correct type for file_type_size.
GitOrigin-RevId: 63baf89ab57485f0bc0f902590271c40f34cf151

+++ td/telegram/files/FileStats.cpp
@@ -44,7 +44,7 @@ void FileStats::add(FullFileInfo &&info) {
 
 FileTypeStat get_nontemp_stat(const FileStats::StatByType &by_type) {
   FileTypeStat stat;
-  for (size_t i = 0; i < file_type_size; i++) {
+  for (int32 i = 0; i < file_type_size; i++) {
     if (FileType(i) != FileType::Temp) {
       stat.size += by_type[i].size;
       stat.cnt += by_type[i].cnt;
@@ -103,7 +103,7 @@ void FileStats::apply_dialog_limit(int32 limit) {
     if (all_dialogs.count(it->first)) {
       ++it;
     } else {
-      for (size_t i = 0; i < file_type_size; i++) {
+      for (int32 i = 0; i < file_type_size; i++) {
         other_stats[i].size += it->second[i].size;
         other_stats[i].cnt += it->second[i].cnt;
       }
@@ -123,7 +123,7 @@ tl_object_ptr<td_api::storageStatisticsByChat> as_td_api(DialogId dialog_id,
   auto stats = make_tl_object<td_api::storageStatisticsByChat>(dialog_id.get(), 0, 0, Auto());
   int64 secure_raw_size = 0;
   int32 secure_raw_cnt = 0;
-  for (size_t i = 0; i < file_type_size; i++) {
+  for (int32 i = 0; i < file_type_size; i++) {
     FileType file_type = static_cast<FileType>(i);
     auto size = stat_by_type[i].size;
     auto cnt = stat_by_type[i].cnt;
@@ -200,7 +200,7 @@ StringBuilder &operator<<(StringBuilder &sb, const FileStats &file_stats) {
     }
 
     sb << "[FileStat " << tag("total", total_stat);
-    for (int i = 0; i < file_type_size; i++) {
+    for (int32 i = 0; i < file_type_size; i++) {
       sb << tag(Slice(file_type_name[i]), file_stats.stat_by_type[i]);
     }
     sb << "]";
@@ -223,7 +223,7 @@ StringBuilder &operator<<(StringBuilder &sb, const FileStats &file_stats) {
       }
 
       sb << "[FileStat " << tag("owner_dialog_id", by_type.first) << tag("total", dialog_stat);
-      for (int i = 0; i < file_type_size; i++) {
+      for (int32 i = 0; i < file_type_size; i++) {
         sb << tag(Slice(file_type_name[i]), by_type.second[i]);
       }
       sb << "]";
